### 2024-01-09 #6198 @editorialbot
[PRE REVIEW]: lbh15: a Python package for standard use and implementation of physical data of heavy liquid metals used in nuclear reactors
**Submitting author:** @lelaus (<a href="http://orcid.org/0000-0003-0937-8897">Gabriele Ottino</a>)
**Repository:** https://github.com/newcleo-dev-team/lbh15
**Branch with paper.md** (empty if default branch): 133_paper_joss
**Version:** v2.0.0
