### 2024-01-10 #6198 @kellyrowland
hi @gwenchee @sskutnik 👋 would you be interested in and available for reviewing this JOSS submission?
