### 2024-02-21 56b170b
[136_single_prop_avail_corrs], issue - #136, Added method to retrieve available correlations for specific properties
