### 2024-02-21 newcleo-dev-team/lbh15#140 @LukeSeifert
JOSS review: Documentation
This issue comes from the JOSS review [here](https://github.com/openjournals/joss-reviews/issues/6383).

### A Statement of Need
I don't think it is immediately clear who the target audience for this software is or what problem it is designed to solve. I would suggest having a sentence or two in the readme and the documentation specifying exactly who should use this software, and why they should use it. This could then also be tied back to the tutorial case given as an example use-case.

### Community Guidelines
The documentation does contain contact information, but I think including a contributing document would also be useful. The three guidelines that should be provided are for third-parties to:

1) Contribute to the software
2) Report issues or problems with the software
3) Seek support
