### 2024-02-21 newcleo-dev-team/lbh15#141 @LukeSeifert
JOSS review: Software Paper
This issue comes from the JOSS review [here](https://github.com/openjournals/joss-reviews/issues/6383).

### Quality of writing
I have a few comments on areas that could be rewritten slightly. If there are disagreements, please feel free to let me know.

#### Summary
- Line 15: "the pressure values" to "pressure values".
- Line 15: "In addition, the value" refers to what value? Unclear if temperature, pressure, or something else.

#### Statement of Need
- Line 20: Replace "It" with "Thermal-hydraulic analysis"
- Line 22: "computer tools" to "computational tools"
- Line 22: "With a growing community..." rewrite to "A standardization of the methods is necessary... results. This standardization is particularly important considering the growing community of users with robust quality assurance needs."
- Line 24: "homogeneity, reproducibility and comparability" missing a comma after reproducibility.
- Line 25: "This is an essential point..." rewrite to "This is an essential point to ensure effective and successful..."
- Line 29: "computer tools" to "computational tools"

#### Implementation
- Line 31: "that" to "which"
- Line 35: "that" to "which"
- Line 37: "implement directly" to "directly implement"
- Line 37: "but it" to "but it instead"

#### Use
- Line 43: remove "that is"
- Lines 43-44: Consider splitting this into multiple sentences
- Line 45: "one single" to "a single"
- Line 46: "after checking its validity" what is "its" referring to here? Unclear if it is the approach or properties.
- Lines 47-49: "In addition..." consider rephrasing this sentence
- Lines 49-52: What does "where it may be too much to know..." mean? Does this approach reduce the number of outputs?

#### Implemented Properties
- Lines 54-61: These two bullet points are not structured in the same way. The first has everything in parentheses, and the second has everything separated by a comma. Consider using a colon followed by listing each item.

#### Implementation History
- Lines 66-68: Instead of "solutions have been adopted", consider removing the parentheses and saying explicitly vectorization and Horner scheme (also I believe it is "vectorization" in American English and "vectorisation" in British English).
- Line 68: Replace "it" with "The current version"
- Line 71: Remove ", see"

#### Documentation
- Lines 82-83: Replace "is completed with" with "contains"
- Line 83: "the users" to "users"
- Line 83: "...starting from basic use up to..." to "...from basic use to..."
