### 2024-02-23 27c1fd5
[136_single_prop_avail_corrs], issue - #136, Updated method available_corrs_for_properties and related docstring
