### 2024-02-23 1bb7f00
[136_single_prop_avail_corrs], issue - #136, Modified type hint into available_corrs_for_property method doc string
