### 2024-02-24 e03bea8
[136_single_prop_avail_corrs], issue - #136, Merged available_corrs_for_property() method into available_correlations()
