### 2024-02-24 2d4d4b5
[136_single_prop_avail_corrs], issue - #136, A few style corrections
