### 2024-02-26 d660d1b
[136_single_prop_avail_corrs], issue - #136, Modified doc for adding example about the new features of available_correlations() method
