### 2024-03-01 newcleo-dev-team/lbh15#144 @damar-wicaksono
JOSS Submission Review (Reviewer 1): Functionality
This issue is related to the [JOSS submission review](https://github.com/openjournals/joss-reviews/issues/6383).

## JOSS Review: Functionality

- [x] Calling `print()` function on an instance raised an exception

```python
>>> from lbh15 import Lead
>>> liquid_lead = Lead(T=750)
>>> print(liquid_lead)
...in LiquidMetalInterface.__str__(self)
    722 for key in self.__properties:
    723     prop_name = key.replace("_"+type(self).__name__, "")
--> 724     info = getattr(self, prop_name+"_info")(print_info=False, n_tab=1)
    725     rvalue += info + "\n"
    726 return rvalue

TypeError: <lambda>() got an unexpected keyword argument 'print_info'
```

While the **Basic Usage** section of the documentation demonstrates the use of `pprint`, I think the default `print()` function should also work for instances, at least, it should not raise an exception.

```python
>>> getattr(liquid_lead, "k_info")
k:
        Value: 17.45 [W/(m*K)]
        Validity range: [600.60, 1300.00] K
        Correlation name: 'lbh15'
        Long name: thermal conductivity
        Units: [W/(m*K)]
        Description:
                Liquid lead thermal conductivity
```

I think the problem is really about adidtional keyword arguments for the lambda function.

- [x] The outcome of `pprint.pprint()` and `__repr__()` is undistinguishable (by design?).

Consider:

```python
>>> import lbh15
>>> import pprint
>>> liquid_lead = lbh15.Lead(T=1000.0)
>>> liquid_lead
C:\Users\...\_lbh15.py: UserWarning: The cobalt diffusivity is requested at temperature value of 1000.00 K that is not in validity range [1023.00, 1273.00] K
  attr_value = getattr(self, key)
C:\Users\...\_lbh15.py:731: UserWarning: The Oxygen partial pressure divided by the oxygen concentration squared is requested at temperature value of 1000.00 K that is not in validity range [783.00, 973.00] K
  attr_value = getattr(self, key)
C:\Users\...\_lbh15.py:731: UserWarning: The nickel solubility is requested at temperature value of 1000.00 K that is not in validity range [598.00, 917.00] K
  attr_value = getattr(self, key)
C:\Users\...\_lbh15.py:731: UserWarning: The silicon solubility is requested at temperature value of 1000.00 K that is not in validity range [1323.00, 1523.00] K
  attr_value = getattr(self, key)
C:\Users\...\_lbh15.py:731: UserWarning: The Oxygen concentration lower limit times nickel concentration is requested at temperature value of 1000.00 K that is not in validity range [673.00, 917.00] K
  attr_value = getattr(self, key)
Lead(T=1000.00, p=101325.00, lim_fe_sat=1.25e-06, lim_fe=5.77e-09, u_s=1707.00, fe_dif=2.48e-05, lim_si_sat=2.28e-15, o_sol=0.02, mu=1.33e-03, p_s=1.41, lim_ni_sat=4.30e-04, se_dif=7.16e-05, cr_sol=9.38e-04, in_dif=5.90e-05, S=15.28, fe_sol=7.67e-04, h=57656.86, G=-3.34e+03, co_dif=3.20e-05, beta_s=3.38e-11, cp=140.89, rho=10161.50, H=11946.50, r=1.14e-06, o_pp=1.03e-09, lim_cr=1.43e-13, sigma=0.41, ni_sol=0.92, o_dif=9.45e-06, lim_si=5.14e-17, si_sol=5.08e-04, k=20.20, te_dif=4.59e-05, lim_al_sat=5.24e-20, alpha=1.26e-04, lim_cr_sat=1.49e-11, lim_ni=3.97e-04)
```

compare with

```python
>>> pprint.pprint(liquid_lead)
C:\Users\...\_lbh15.py:731: UserWarning: The cobalt diffusivity is requested at temperature value of 1000.00 K that is not in validity range [1023.00, 1273.00] K
  attr_value = getattr(self, key)
C:\Users\...\_lbh15.py:731: UserWarning: The Oxygen partial pressure divided by the oxygen concentration squared is requested at temperature value of 1000.00 K that is not in validity range [783.00, 973.00] K
  attr_value = getattr(self, key)
C:\Users\...\_lbh15.py:731: UserWarning: The nickel solubility is requested at temperature value of 1000.00 K that is not in validity range [598.00, 917.00] K
  attr_value = getattr(self, key)
C:\Users\...\_lbh15.py:731: UserWarning: The silicon solubility is requested at temperature value of 1000.00 K that is not in validity range [1323.00, 1523.00] K
  attr_value = getattr(self, key)
C:\Users\...\_lbh15.py:731: UserWarning: The Oxygen concentration lower limit times nickel concentration is requested at temperature value of 1000.00 K that is not in validity range [673.00, 917.00] K
  attr_value = getattr(self, key)
Lead(T=1000.00, p=101325.00, lim_fe_sat=1.25e-06, lim_fe=5.77e-09, u_s=1707.00, fe_dif=2.48e-05, lim_si_sat=2.28e-15, o_sol=0.02, mu=1.33e-03, p_s=1.41, lim_ni_sat=4.30e-04, se_dif=7.16e-05, cr_sol=9.38e-04, in_dif=5.90e-05, S=15.28, fe_sol=7.67e-04, h=57656.86, G=-3.34e+03, co_dif=3.20e-05, beta_s=3.38e-11, cp=140.89, rho=10161.50, H=11946.50, r=1.14e-06, o_pp=1.03e-09, lim_cr=1.43e-13, sigma=0.41, ni_sol=0.92, o_dif=9.45e-06, lim_si=5.14e-17, si_sol=5.08e-04, k=20.20, te_dif=4.59e-05, lim_al_sat=5.24e-20, alpha=1.26e-04, lim_cr_sat=1.49e-11, lim_ni=3.97e-04)
```

Is there any reason to use `pprint` at all?

- [x] Asking for the available correlations for a few specific properties seems to raise an exception:

```python
>>> from lbh15 import Lead
>>> Lead.available_correlations(["cp", "u_s"])
Cell In[88], line 1
----> 1 Lead.available_correlations(["cp", "u_s"])

TypeError: available_correlations() takes 1 positional argument but 2 were given
```

Considering Issue #136, I assume the version in PyPI (that I used for reviewing the package) does not support additional argument, yet?

- [x] `correlations_available()` or `available_correlations()`.

In the advanced usage there are two method names as follows.

```python
>>> Lead.available_correlations()
defaultdict(<class 'list'>, {'k': ['lbh15'], 'lim_ni': ['lbh15'], 'u_s': ['sobolev2011'], 'mu': ['lbh15'], 'H': ['lbh15'], 'lim_cr': ['venkatraman1988', 'alden1958', 'gosse2014'], 'sigma': ['jauch1986'], 'cp': ['sobolev2011', 'gurvich1991'], 'cr_sol': ['venkatraman1988', 'gosse2014', 'alden1958'], 'in_dif': ['lbh15'], 'o_pp': ['taskinen1979', 'charle1976', 'alcock1964', 'otsuka1979', 'otsuka1981', 'fisher1966', 'isecke1977', 'szwarc1972', 'ganesan2006'], 'lim_cr_sat': ['lbh15'], 'si_sol': ['lbh15'], 'fe_dif': ['lbh15'], 'lim_si': ['lbh15'], 'o_dif': ['charle1976', 'swzarc1972', 'arcella1968', 'gromov1996', 'ganesan2006b', 'otsuka1975', 'homna1971'], 'G': ['lbh15'], 'p_s': ['sobolev2011'], 'beta_s': ['lbh15'], 'se_dif': ['lbh15'], 'lim_al_sat': ['lbh15'], 'ni_sol': ['gosse2014'], 'r': ['lbh15'], 'S': ['lbh15'], 'h': ['sobolev2011'], 'alpha': ['lbh15'], 'co_dif': ['lbh15'], 'lim_si_sat': ['lbh15'], 'lim_fe_sat': ['lbh15'], 'fe_sol': ['gosse2014'], 'lim_fe': ['lbh15'], 'rho': ['sobolev2008a'], 'te_dif': ['lbh15'], 'lim_ni_sat': ['lbh15'], 'o_sol': ['lbh15']})
```

and

```python
>>> Lead.correlations_available()
defaultdict(<class 'list'>, {'o_pp': ['taskinen1979', 'charle1976', 'alcock1964', 'otsuka1979', 'otsuka1981', 'fisher1966', 'isecke1977', 'szwarc1972', 'ganesan2006'], 'lim_cr_sat': ['lbh15'], 'fe_dif': ['lbh15'], 'o_dif': ['charle1976', 'swzarc1972', 'arcella1968', 'gromov1996', 'ganesan2006b', 'otsuka1975', 'homna1971'], 'rho': ['sobolev2008a', 'custom2022'], 'cp': ['gurvich1991', 'sobolev2011'], 'lim_si': ['lbh15'], 'G': ['lbh15'], 'se_dif': ['lbh15'], 'ni_sol': ['gosse2014'], 'lim_al_sat': ['lbh15'], 'k': ['lbh15'], 'si_sol': ['lbh15'], 'u_s': ['sobolev2011'], 'mu': ['lbh15'], 'sigma': ['jauch1986'], 'cr_sol': ['gosse2014', 'alden1958', 'venkatraman1988'], 'S': ['lbh15'], 'lim_cr': ['alden1958', 'gosse2014', 'venkatraman1988'], 'H': ['lbh15'], 'co_dif': ['lbh15'], 'fe_sol': ['gosse2014'], 'lim_si_sat': ['lbh15'], 'lim_fe_sat': ['lbh15'], 'lim_fe': ['lbh15'], 'beta_s': ['lbh15'], 'p_s': ['sobolev2011'], 'te_dif': ['lbh15'], 'lim_ni_sat': ['lbh15'], 'o_sol': ['lbh15'], 'r': ['lbh15'], 'lim_ni': ['lbh15'], 'h': ['sobolev2011'], 'alpha': ['lbh15'], 'in_dif': ['lbh15'], 'T_double': ['double2022']})
```

Is there any difference? Note that the package available in PyPI does not recognize `correlations_available()`.

- [x] (Possible edge case): Path for module implementing custom properties are not checked for invalid separator when added.

For instance, the example (Advanced Usage):

```python
...
>>> Lead.set_custom_properties_path(os.getcwd() + '/custom_properties/lead_properties.py')
```

is invalid for Windows due to the separator so calling:

```python
>>> Lead.available_correlations()
...
ModuleNotFoundError: No module named ...
```

Now suppose I have correctly specified the full path of the module the first time (in Windows):

```python
>>> Lead.set_custom_properties_path(os.getcwd() + "\\custom_properties\\lead_properties.py")
>>> Lead.available_correlations()
defaultdict(list,
            {'h': ['sobolev2011'],
             'lim_al_sat': ['lbh15'],
             ...
             'rho': ['sobolev2008a', 'custom2022'],
             ...
```

If I now add the custom module the second time incorrectly (invalid separators):

```python
>>> Lead.set_custom_properties_path(os.getcwd() + "/custom_properties/lead_properties.py")
>>> Lead.available_correlations()
...
ModuleNotFoundError: No module named ...
```

This because the second module is incorrectly specified but not caught when added.
If there is no such file in `<exec_dir>/custom_properties` there is a safeguard; here, however, the file exists but later on incorrectly specified for `importlib`.

Furthermore, there is no official way to rectify this because the custom paths are stored in a private attribute (perhaps, I have to delete the key manually to correct this):

```python
>>> Lead._custom_properties_path
{'C:\\Users\\HOME\\projects\\lbh15-review\\custom_properties\\': ['lead_properties',
  'double_T'],
 'C:\\Users\\HOME\\projects\\': ['lbh15-review/test_properties/lead_properties']}  # This is invalid
```

I understand that this is most probably an edge case but it can be confusing
because `set_custom_properties_path()` appends an internal dictionary. If one of the path is invalid but somehow uncaught, an exception will be raised when the properties are listed.
Perhaps you can use a path normalization feature of either `pathlib` or `os` to deal with such a cross-platform issue?
