### 2024-03-01 newcleo-dev-team/lbh15#145 @damar-wicaksono
JOSS Submission Review (Reviewer 1): Software Paper
This issue is related to the [JOSS submission review](https://github.com/openjournals/joss-reviews/issues/6383).

Please find below some questions, remarks, and suggestions regarding the draft paper. Please feel free to let me know if you need clarification. Thanks!

**Summary**

- Line 14 "_...defined by its thermo-dynamic state,..._": I think it's more common to write "thermodynamic" instead of "thermo-dynamic".
- Lines 15-16 "_In addition, the value of several other properties can be used at liquid metal object’s instantiation,..._": This is not clear. Which other properties? Furthermore, I think (because I follow the tutorials) it means that the value of these properties can be used _instead_ of the temperature and pressure to instantiate the object (with the given injectivity condition); if so, this must be stated clearly here.

**Statement of Need**:

- Lines 28-29 "_In this context, standard libraries providing the physical correlations for thermal-hydraulic computer tools are needed._": It is not clear from this supposedly prominent sentence what "physical correlations" actually refers to (only in connection with the summary it is clear; by the way, I think they should be correlations of physical properties). Furthermore, "Standard libraries of physical correlations for thermal-hydraulic computational tools" are very broad and I think the closing of a statement of need should be a bit more focused.
- Lines 25-26, "_...in both industry and research environment,..._": Perhaps "industrial" instead of "industry" would be more apt?

**Implementation**:

- Line 31: The package `iapws` cited by the paper is archived at Zenodo with its own [DOI:10.5281/zenodo.4744318](https://zenodo.org/records/4744318). Perhaps this should appear in the list of references instead of a footnote.
- Line 34 "_The efficiency and the effectiveness are assured by the Object-Oriented ... approaches,..._": It may be more common to say "object-oriented design" instead of "object-oriented approach".

**Use**:

- Line 46: "thermodynamic" instead of "thermo-dynamic".
- Line 48 "_In addition, the former approach allows to manage the properties correlations in a simpler way..._": I think the verb "allows + to" is transitive, it needs an object, perhaps "the former approach allows users/one to manage...".
- Line 48 "_...by acting either on a single or on all liquid metal instances at the same time._": I'm not sure I understand the part "on all liquid metal instances at the same time" written here; If I instantiate _a_ liquid metal object to access its related properties, how can I act on all instances? Perhaps I misunderstood it.
- Line 51 "_...where it may be too much to know the values of all the properties..._": In which sense it is "too much" to know here? Is it in the sense of "knowing the values of all the properties are _unnecessary_"? Or?

**Implemented Properties**:

- These two subdivisions are not structured parallely, either use parentheses or "..., including...".
- I see no reason to capitalize the term "Oxygen".

**Implementation History**:

- Line 66, "_It improves the performance and the readability._": The readability of what? The documentation or the codebase as a whole?
- Line 67, "_... improving performance and usability (..._": In what way using vectorization and Horner's scheme improve usability of the package?
- Line 68, "_It includes a tutorial focusing on a volume of lead that is subjected to time-varying thermal loads..._": The sentence that came before focused on improved implementations, but this sentence focused on the documentation. It should be stated explicitly instead of using the pronoun "it".
- Lines 71-72, "_Great attention is paid to the code quality and readability_": This is somehow a repetition and extension of the second sentence of the paragraph. Perhaps combine them into a single sentence?
- Maybe add a footnote for the repository of `pycodestyle` and `pylint`?
- Lines 72-74, "_In addition, the automatic static analysis of the code has been performed by applying the pylint tool throughout the entire development, achieving a score higher than 9 out of 10._": Is stating the score really necessary here? While it is a very high score, it would be difficult to judge this number without really understanding what is being analyzed by `pylint`? Would it be enough to state that you ensure the quality of the code using tools like `pycodestyle` for styling and `pylint` for linting (and static code analysis)?

**Documentation**:

- Line 82, "_An advanced use of the package needs skills in Python software development_.": This is a very vague statement, what kind of skills in Python software development, specifically? Perhaps more importantly, if such a sentence is necessary, what is considered an "advanced use" of the package?
- Line 83, "_The documentation is completed with examples..._": Is the documentation complete? Or it will be kept getting updated along with new versions of the package? Perhaps "also includes various examples" would suffice?
- Line 83, "_...for the users, starting from basic use up to short tutorials for more advanced applications._":
  - "users" instead of "the users"
  - "starting from...": "starting" is not necessary
  - the term adopted in the documentation is "Basic Usage" instead of "Basic Use".
  - Similarly is it "advanced applications" or "advanced usage"?
