### 2024-03-04 99b6b06
[137_choose_modules], issue - #137, Transformed the list of modules to load into a dictionary
