+++ lbh15/lead.py
@@ -98,12 +98,16 @@ class Lead(LiquidMetalInterface):
     _custom_properties_path: Dict[str, List[str]] = {}
     _available_properties_dict: Dict[str, PropertyInterface] = {}
     _available_correlations_dict: Dict[str, List[str]] = {}
-    _properties_modules_list: List[str] = \
-        ['lbh15.properties.lead_thermochemical_properties.solubility_in_lead',
-         'lbh15.properties.lead_thermochemical_properties.diffusivity_in_lead',
-         'lbh15.properties.lead_thermochemical_properties.lead_thermochemical',
-         'lbh15.properties.lead_thermochemical_properties.lead_oxygen_limits',
-         'lbh15.properties.lead_properties']
+    _properties_modules_dict: Dict[str, str] = \
+        {'solubility': 'lbh15.properties.lead_thermochemical_properties\
+.solubility_in_lead',
+         'diffusivity': 'lbh15.properties.lead_thermochemical_properties\
+.diffusivity_in_lead',
+         'thermochem': 'lbh15.properties.lead_thermochemical_properties\
+.lead_thermochemical',
+         'oxygenlims': 'lbh15.properties.lead_thermochemical_properties\
+.lead_oxygen_limits',
+         'thermophys': 'lbh15.properties.lead_properties'}
 
     def __init__(self, p: float = atm, **kwargs):
         self._guess = LEAD_BOILING_TEMPERATURE / 2.0
