### 2024-03-04 newcleo-dev-team/lbh15#147 @damar-wicaksono
JOSS Submission Review (Reviewer 1): Documentation
This issue is related to the [JOSS submission review](https://github.com/openjournals/joss-reviews/issues/6383).

Please find below some remarks, questions, and suggestions regarding the documentation and testing of the package; just let me know if clarifications are needed.

As a general remark, please be consistent with the choice of English spelling, either American or British both in the paper and the docs. I see a couple of mix-up, e.g., "behavior", "vapour".

### Introduction

- [ ] Add a statement of need in README and in the documentation (either in the landing page or Introduction):
  - This statement may be a summary of the one appears in the paper; give one- or two-line summary regarding the purpose of the package, its target audience, and its intended (common) use.
  - Currently the **Introduction** section only answers "what is `lbh15`".

- [ ] (Suggestion) add the additional requirements to build the docs and testing for prospective contributors in the toml/`setup.py` file.

### Basic Usage:

- [ ] The summary information of the liquid metal using *pprint* package looks the same as `__repr__()` (by design)?
  -  as already mentioned in the remarks on `Functionality`

### Initialization from Properties (_experimental_)

- [ ] "_Such function, which constitutes a physical correlation, must be injective in the range considered for the root._": Perhaps it would be more apt to say "physical property correlation" instead of "physical correlation"?

### Advanced Usage:

- [ ] The whole section should be reorganized into two subsections: Adding a new correlation of a given property _and_ Adding a new property (with its own correlation).
- [ ] The first bullet point can be reformulated focusing instead on introducing the function to query the available correlations either for all physical properties or selected properties. Then, the next bullet point can focus on introducing a new property implemented as a Pyton module.
- [ ] "_Let implement the new property in..._": I believe it should be "Let's" instead of "Let".
- [ ] As noted in the Functionality review, there should either be a note on the path separator in different OS (particularly Windows) or make the internal implementation agnostic regarding the separator.
- [ ] Instead of showing all the available correlations after adding a new one, perhaps focus on listing the relevant correlation, i.e., `rho`.

  >>> from lbh15 import Lead
  >>> import os
  >>> Lead.set_custom_properties_path(os.getcwd() + '/custom_properties/lead_properties.py')
  >>> Lead.correlations_available()
  ...
  ```

- [ ] Is it `Lead.available_correlations()` or `Lead.correlations_available()`?
- [ ] "_If the density correlation is not specified for a new object instantiation, the last one in the list will be selected as default:_": The example below should also shows the default behavior as noted here if the function `Lead.set_correlation_to_use()` is not used; should it automatically use `custom2022` as it is the last one in the list?
- [ ] "_For instance, let’s implement in `<execution_dir>/custom_property/double_T.py`_": Note that in the preceeding discussion the custom properties are stored in `<execution_dir>/custom_properties`.

### Learn More

- [ ] "_...describes a tutorial application coming together with *lbh15*,..._": Consider rephrasing "coming together with" with "delivered with".
- [ ] Is there any particular reason to capitalize "Oxygen" in most places? As a common noun, I'm not sure that would be necessary.

### Oxygen Control

- [ ] "_Oxygen is the most important chemical element, which results from start-up operations, maintenance services and, **possibily**, incidental contaminations..._": Typo "possibily", should have been "possibly".

**Oxygen Concentration Lower Limit**:

- [ ] Point 4, "_Ongoing researches are in progress, but currently the exact values for the chemical activities of the dissolved metal and of the Oxygen are not known._": The plural form "researches" is uncommon; simply use "research". Furthermore, "ongoing research are in progress" is redundant, it's enough to say that the relevant research is "on going" or "in progress".

**Ranges of Validity**:

- [ ] "_The temperature range of validity specified in the *lbh15* package for each correlation is the most restrictive one._": In what sense the range is the "most restrictive one"?
- [ ] Consistent use of temperature unit. For instance, "673;1000 *K* range" while in the previous section unit is always written as $[K]$.
- [ ] Are multiple additional sources in this section, e.g., tecdoc (2002) also from the Handbook? If not, I think they should be listed in the bibliography of the documentation.
- [ ] "_the correlations are preferred whose validity ranges are related to the lowest available temperature values and whose extension is the widest available._": What does it mean here "extension"? Does it mean "extrapolation" or perhaps something else?

### Tutorials

- [ ] "_...like an Heaviside function..._": "like **a** heaviside function".
- [ ] "_Let suppose that the lead..._": As a conjunction, the sentence can start with "Suppose the lead...".
- [ ] "_...an Iron..._": I don't think "Iron" should be capitalized.
- [ ] I think the tutorial script also requires `matplotlib` which unlike `simple-pid` is not listed.

### API Guide

- [ ]  Please double check the following two URL's: https://newcleo-dev-team.github.io/lbh15/source/documentation.html and (by clicking next) https://newcleo-dev-team.github.io/lbh15/source/lbh15.html. Are both necessary?

- [ ] **lead Module**, **bismuth Module**, and **lbe Module**:
  - [ ] "_A *lead* instance..._": Should it be "`Lead` instance" (respectively, `Bismuth` and `LBE`)?
  - [ ] I think the listed thermo-physical and thermo-chemical properties in https://newcleo-dev-team.github.io/lbh15/source/lead.html should have been written (or additionally) in monospace as they are directly related to properties of an instance.
  - [ ] About thermo-chemical property modules for each materials: Is there any particular reason to re-implement `OxygenDiffusivityInterface` class in each module? A similar question for `IronSolubility`.

There is no section in the docs regarding contribution guidelines. I believe this is a requirement of JOSS, so please consider adding them following the common best-practices. It should include, according to the reviewer guideline:

- Contribute to the software
- Report issues or problems with the software
- Seek support

Please also include the protocol for testing the package. In particular, which framework does it employ (`pytest`? `unittest`?)?

### Regarding the tests:

I tried to run the tests of the package from the repo. I'm not sure If I did everything correctly, but I got these errors:

```
FAILED test_bismuth_fromX_spanT.py::BismuthTester::test_init_fromX - AssertionError: 548.15 != 1102.673571428572 within 8 places (554.5235714285719 difference) : rho FAILED
FAILED test_lead_basics.py::LeadComparisonTester::test_vs_data - AssertionError: 10652.22 != 10643.67958 within 4 places (8.540419999999358 difference)
FAILED test_lead_fromX_spanT.py::LeadTester::test_init_fromX - AssertionError: 1573.15 != 1794.2094619326822 within 8 places (221.0594619326821 difference) : cp FAILED
```

Perhaps you could double check them? Thanks a lot!
