### 2024-03-04 1c4ea01
[148_statement_of_need], issue - #148, Improved statement of need description into doc
