### 2024-03-05 e8a139e
[150_contribution], issue - #150, Added files devoted to contribution and code of conduct guidelines
