+++ CODE_OF_CONDUCT.rst
@@ -1,0 +1,103 @@
+Contributor Code of Conduct
+===========================
+
+Our Pledge
+----------
+
+We as members, contributors, and leaders pledge to make participation in our
+community a harassment-free experience for everyone, regardless of age, body
+size, visible or invisible disability, ethnicity, sex characteristics, gender
+identity and expression, level of experience, education, socio-economic status,
+nationality, personal appearance, race, caste, color, religion, or sexual
+identity and orientation.
+We pledge to act and interact in ways that contribute to an open, welcoming,
+diverse, inclusive, and healthy community.
+
+Our Standards
+-------------
+
+Examples of behavior that contributes to a positive environment for our
+community include:
+
+- Demonstrating empathy and kindness toward other people;
+- Being respectful of differing opinions, viewpoints, and experiences;
+- Giving and gracefully accepting constructive feedback;
+- Accepting responsibility and apologizing to those affected by our mistakes, and learning from the experience;
+- Focusing on what is best not just for us as individuals, but for the overall community.
+
+Examples of unacceptable behavior include:
+
+- The use of sexualized language or imagery, and sexual attention or advances of any kind;
+- Trolling, insulting or derogatory comments, and personal or political attacks;
+- Public or private harassment;
+- Publishing others' private information, such as a physical or email address, without their explicit permission;
+- Other conduct which could reasonably be considered inappropriate in a professional setting.
+
+Enforcement Responsibilities
+----------------------------
+
+Project's owners and mantainers are responsible for clarifying and enforcing
+our standards of acceptable behavior and will take appropriate and fair
+corrective action in response to any behavior that they deem inappropriate,
+threatening, offensive, or harmful.
+Project's owners and mantainers have the right and responsibility to remove,
+edit, or reject comments, commits, code, issues, and other contributions that
+are not aligned to this Code of Conduct, and will communicate reasons for
+moderation decisions when appropriate.
+
+Scope
+-----
+
+This Code of Conduct applies within this project's repository, and also applies when
+an individual is officially representing this project in public spaces.
+
+Enforcement
+-----------
+
+Instances of abusive, harassing, or otherwise unacceptable behavior may be
+reported to the project's owners and mantainers by emailing them.
+All complaints will be reviewed and investigated promptly and fairly.
+All project's owners and mantainers are required to respect the privacy and
+security of the reporter of any incident.
+
+Project's Repository Impact Guidelines
+--------------------------------------
+
+Project's owners and mantainers will follow these Project's Repository Impact Guidelines
+in determining the consequences for any action they deem in violation of this
+Code of Conduct:
+
+1. **Correction**
+
+    | **Project Impact**: Use of inappropriate language or other behavior deemed unprofessional or unwelcome in the project's repository.
+    | **Consequence**: A private, written warning from project's owners and mantainers, providing clarity around the nature of the violation and an explanation of why the behavior was inappropriate. A public apology may be requested.
+
+2. **Warning**
+
+    | **Project Impact**: A violation through a single incident or series of actions.
+    | **Consequence**: A warning with consequences for continued behavior. No interaction with the people involved, including unsolicited interaction with those enforcing the Code of Conduct, for a specified period of time. This includes avoiding interactions in community spaces as well as external channels like social media. Violating these terms may lead to a temporary or permanent ban.
+
+3. **Temporary Ban**
+
+    | **Project Impact**: A serious violation of project's repository standards, including sustained inappropriate behavior.
[…]
