+++ CONTRIBUTIONS.rst
@@ -1,0 +1,39 @@
+Contributing guidelines
+=======================
+
+| This guide gives an overview of the project's contribution workflow.
+| Thank you from the outset for your interest and time in contributing to the project!
+
+Issues and feature requests
+---------------------------
+
+| If you find a problem with the project or if you have a feature request, you can open a new issue by clicking on the *New issue* button in the top menu bar at this `web page <https://github.com/newcleo-dev-team/lbh15/issues>`_.
+| If you find an issue or feature request at this `web page <https://github.com/newcleo-dev-team/lbh15/issues>`_ that you are interested in and would like to help solve or implement, please contact the authors by emailing them.
+
+Contribution process
+--------------------
+
+| **Authors acceptance**: Your request of contribution will be carefully analysed by the authors of *lbh15*; upon their approval, you will be added to the developers' group;
+| **Clone the Project**: Next, clone the project to your local machine. This allows you to make changes to the code;
+| **Create a New Branch**: Before making any change, create a new branch in your local project. This helps isolate your changes and makes it easier to submit a merge request later. Use a descriptive name for your branch: the name of the branch is suggested to start with the index that identifies the issue you will be working on.
+| **Make Your Changes**: Now you're ready to make your changes. Make sure you follow the project's coding standards and guidelines.
+| **Commit Your Changes**: After making your changes, commit them to your local project. Write a clear and concise commit message describing what you've done.
+| **Push Your Changes**: Push your changes and the new branch to the remote project.
+| **Submit a Merge Request**: Finally, navigate to the `project <https://github.com/newcleo-dev-team/lbh15>`_ on GitHub and click on the *Pull request* button in the top menu bar. Fill out the form and submit your merge request. In case of merge conflicts, please try to reduce them as much as possibile.
+| 
+| **Remember**: Pull regularly from the original project to keep your local copy up to date. This helps avoid merge conflicts.
+
+Review process
+--------------
+
+Once you've submitted a merge request, it goes through a review process. Here's what to expect:
+
+| **Automated Checks**: After submission, automated checks will run to ensure all tests pass with your changes.
+| **Maintainer Review**: The project's maintainers will review your merge request. They may provide feedback and request changes. Please, respond to comments and make the requested changes, if any.
+| **Revision**: If changes are requested, make the changes in your branch and commit them. GitHub will automatically update the merge request with your new commits.
+| **Approval**: Once the maintainers are satisfied with the changes, they will approve the merge request.
+| **Merge**: After approval, the maintainers will merge your changes into the main project. Your contribution is now part of the project!
+| 
+| **Remember**: The review process can take time. Please, be patient and respectful of the maintainers' time.
+
+These contributing guidelines have been adapted from https://gitlab.com/jtagusari/hrisk-noisemodelling.
