### 2024-03-07 d2e2b60
[141_joss_paper_reviewer_2], issue - #141, Applied modifications to the paper according to comments of reviewer 2
