+++ joss/paper.md
@@ -30,50 +30,55 @@ bibliography: paper.bib
 
 # Summary
 
-`lbh15` is a Python package that provides function correlations for the physical 
-properties of the liquid metals used as coolant in GEN-IV liquid metal fast 
+`lbh15` is a Python package that provides function correlations for the 
+physical properties of the liquid metals used as coolant in GEN-IV liquid metal fast 
 reactors (*LMFR*), such as those cooled by molten lead and lead-bismuth 
 eutectic alloy. The package implements the correlations contained in the 
 reference handbook edited by OECD/NEA [@nea], also offering the possibility of 
 adding new customized properties with minimal effort for the user. The 
 properties of the liquid metal are uniquely defined by its thermo-dynamic 
-state, namely by the temperature and the pressure values. In addition, the 
-value of several other properties can be used at liquid metal object's 
-instantiation, provided that the inverse of the corresponding correlation 
-has at least one root in the validity range (*injective function* property).
+state, namely by the temperature and pressure values. As alternative, the 
+physical properties can be used at liquid metal object's instantiation, 
+provided that the inverse of the corresponding correlation has at least one 
+root in the validity range (*injective function* property).
 
 `lbh15` package is released under the *GNU Lesser General Public License v3.0*.
 
 # Statement of Need
 
 Thermal-hydraulic analysis is a key factor for the design and safety studies 
-of *LMFR*s. It involves the implementation and use of several numerical 
-methods and physical data that are employed in different computer tools. 
+of *LMFR*s, involving the implementation and use of several numerical 
+methods and physical data that are employed in different computational tools. 
+A standardization of the methods is necessary to guarantee homogeneity, 
+reproducibility, and comparability of the numerical results. This 
+standardization is particularly important considering the growing community 
+of users with robust quality assurance needs.
+
 With a growing community of users and considering quality assurance needs 
 for the engineering studies, a standardization of the approaches becomes 
 necessary to guarantee homogeneity, reproducibility and comparability of 
-the numerical results. This is an essential point to ensure work 
-effectiveness and successful projects in both industry and research 
+the numerical results. This is an essential point to ensure effective and 
+successful projects in both industry and research 
 environments, especially for nuclear science and engineering. As well, 
 *new*cleo pursues efforts for data standardization to develop new units 
 of lead-cooled fast reactors (*LFR*).
 
 In this context, standard libraries providing the physical correlations for 
-thermal-hydraulic computer tools are needed.
+thermal-hydraulic computational tools are needed.
 
 # Implementation
 
 `lbh15` package takes inspiration from the `iapws`[^1] Python package, 
-that implements the water-related *IAPWS* full standard [@iapws]. However, 
+which implements the water-related *IAPWS* full standard [@iapws]. However, 
 `lbh15` follows a different implementation approach.
 
 The efficiency and the effectiveness are assured by the *Object-Oriented* 
-and the *Dynamic Loading* approaches, that have been applied throughout 
+and the *Dynamic Loading* approaches, which have been applied throughout 
 the entire development process. `lbh15` relies on the abstract liquid metal 
 class: all classes describing the different metals inherit from it. The 
-abstract class does not implement directly the property correlations, but 
-it instantiates the property objects and provides the property values. In 
-other words, the abstract liquid metal class acts as both *factory* of 
+abstract class does not directly implement the property correlations, but 
+it instead instantiates the property objects and provides the property values. 
+In other words, the abstract liquid metal class acts as both *factory* of 
 the property objects and *proxy* of the property values [@pybook]. 
 This allows the user to add new custom properties without modifying the 
 existing implementation of the liquid metal class.
[…]
