### 2024-03-09 bd4066b
[153_check_separator], issue #153, use os.path.normpath to normalize pathname in set_custom_properties_path + used pathlib.os.sep to split the provided path
