### 2024-03-10 b22e5ab
[152_fix_print], issue #152, added arguments to lambda function defined in __add_property
