### 2024-03-10 b250612
[152_fix_print], issue #152, added default values to lambda function
