+++ lbh15/_lbh15.py
@@ -556,7 +556,7 @@ class LiquidMetalInterface(ABC):
         key = property_object.name
         self.__properties[key] = property_object
         setattr(self, property_object.name+"_info",
-                lambda print_info, n_tab:
+                lambda print_info=True, n_tab=0:
                     self.__properties[key].info(self.__T, self.__p,
                                                 print_info, n_tab))
 
