### 2024-03-10 a781bc3
[160_common_interface], fixed f-string in description of some properties
