### 2024-03-10 4a9db76
[160_common_interface], issue #160, defined common OxygenPartialPressureInterface
