### 2024-03-10 90ca50d
[160_common_interface], issue #160, defined common IronSolubilityInterface
