### 2024-03-10 7772a0f
[160_common_interface], issue #160, defined common NickelSolubilityInterface
