### 2024-03-10 4a6306a
[160_common_interface], issue #160, defined common ChromiumSolubilityInterface
