### 2024-03-10 983f3ce
[160_common_interface], issue #160, defined common OxygenSolubilityInterface
