### 2024-03-10 3cb55c8
[160_common_interface], issue #160, defined common MolarEntropyInterface
