+++ lbh15/properties/lbe_thermochemical_properties/lbe_thermochemical.py
@@ -8,6 +8,7 @@ from scipy.constants import R
 from ..interface import PropertyInterface
 from ..common_interface import OxygenPartialPressureInterface
 from ..common_interface import MolarEnthalpyInterface
+from ..common_interface import MolarEntropyInterface
 from ..lbe_properties import h
 from ..._commons import LBE_BOILING_TEMPERATURE as T_b0
 from ..._commons import LBE_MELTING_TEMPERATURE as T_m0
@@ -280,7 +281,7 @@ class MolarEnthalpy(MolarEnthalpyInterface):
         return f"{self.long_name} in liquid lbe"
 
 
-class MolarEntropy(PropertyInterface):
+class MolarEntropy(MolarEntropyInterface):
     """
     Liquid lbe *molar entropy variation* property class.
     """
@@ -313,33 +314,12 @@ class MolarEntropy(PropertyInterface):
                            + 2.28e5 * (1 / T / T - 1 / T_m0 / T_m0))
 
     @property
-    def name(self) -> str:
-        """
-        str : Name of the property
-        """
-        return "S"
-
-    @property
-    def units(self) -> str:
-        """
-        str : Molar entropy variation unit
-        """
-        return "[J/(mol.K)]"
-
-    @property
     def range(self) -> List[float]:
         """
         List[float] : Temperature validity range of the molar entropy
         variation correlation function
         """
         return [400, T_b0]
-
-    @property
-    def long_name(self) -> str:
-        """
-        str : Molar entropy variation long name
-        """
-        return "molar entropy variation"
 
     @property
     def description(self) -> str:
