### 2024-03-10 df0b99b
[160_common_interface], issue #160, defined common GibbsFreeEnergyInterface
