+++ lbh15/properties/lbe_thermochemical_properties/lbe_oxygen_limits.py
@@ -5,7 +5,14 @@ from typing import Union
 import numpy as np
 from scipy.constants import atm
 from scipy.constants import R
-from ..interface import PropertyInterface
+from ..common_interface import LowerLimitSaturationIronInterface
+from ..common_interface import LowerLimitSaturationChromiumInterface
+from ..common_interface import LowerLimitSaturationNickelInterface
+from ..common_interface import LowerLimitSaturationSiliconInterface
+from ..common_interface import LowerLimitSaturationAluminiumInterface
+from ..common_interface import LowerLimitChromiumInterface
+from ..common_interface import LowerLimitNickelInterface
+from ..common_interface import LowerLimitIronInterface
 from .solubility_in_lbe import OxygenSolubility
 from .solubility_in_lbe import ChromiumSolubilityGosse2014
 from .solubility_in_lbe import ChromiumSolubilityCourouau2004
@@ -18,7 +25,7 @@ from .lbe_thermochemical import LeadChemicalActivity
 from ..._decorators import range_warning
 
 
-class LowerLimitSaturationIron(PropertyInterface):
+class LowerLimitSaturationIron(LowerLimitSaturationIronInterface):
     """
     *Lower limit of Oxygen concentration* to promote a
     protective oxide film in liquid lbe considering
@@ -78,34 +85,12 @@ class LowerLimitSaturationIron(PropertyInterface):
             * LeadChemicalActivity().correlation(T, p)
 
     @property
-    def name(self) -> str:
-        """
-        str : Name of the property
-        """
-        return "lim_fe_sat"
-
-    @property
-    def units(self) -> str:
-        """
-        str : Oxygen concentration lower limit unit
-        """
-        return "[wt.%]"
-
-    @property
     def range(self) -> List[float]:
         """
         List[float] : Temperature validity range of the Oxygen concentration
         lower limit correlation function
         """
         return [673, 1000]
-
-    @property
-    def long_name(self) -> str:
-        """
-        str : Oxygen concentration lower limit long name
-        """
-        return ("Oxygen concentration lower limit for"
-                "iron at its saturation concentration")
 
     @property
     def description(self) -> str:
@@ -115,7 +100,7 @@ class LowerLimitSaturationIron(PropertyInterface):
         return f"{self.long_name} in liquid lbe"
 
 
-class LowerLimitSaturationChromium(PropertyInterface):
+class LowerLimitSaturationChromium(LowerLimitSaturationChromiumInterface):
     """
     *Lower limit of Oxygen concentration* to promote a
     protective oxide film in liquid lbe considering
@@ -175,34 +160,12 @@ class LowerLimitSaturationChromium(PropertyInterface):
             * LeadChemicalActivity().correlation(T, p)
 
     @property
-    def name(self) -> str:
-        """
-        str : Name of the property
-        """
-        return "lim_cr_sat"
-
-    @property
-    def units(self) -> str:
-        """
-        str : Oxygen concentration lower limit unit
-        """
-        return "[wt.%]"
-
-    @property
     def range(self) -> List[float]:
         """
         List[float] : Temperature validity range of the Oxygen concentration
         lower limit correlation function
         """
         return [673, 1000]
-
-    @property
-    def long_name(self) -> str:
-        """
-        str : Oxygen concentration lower limit long name
-        """
-        return ("Oxygen concentration lower limit for"
-                "chromium at its saturation concentration")
 
     @property
     def description(self) -> str:
@@ -212,7 +175,7 @@ class LowerLimitSaturationChromium(PropertyInterface):
         return f"{self.long_name} in lbe"
 
 
[…]
