### 2024-03-10 98a5464
[160_common_interface], issue #160, defined common_interface.rst and included in the properties.rst
