+++ docs/source/common_interface.rst
@@ -1,0 +1,7 @@
+*common_interface* Module
+=========================
+
+.. automodule:: lbh15.properties.common_interface
+    :members:
+    :member-order: bysource
+    :inherited-members:
