### 2024-03-10 59453d5
[160_common_interface], issue #160, minor fixes to fit pycodestyle
