+++ lbh15/properties/lead_thermochemical_properties/lead_oxygen_limits.py
@@ -6,14 +6,14 @@ import numpy as np
 from scipy.constants import atm
 from scipy.constants import R
 from ..interface import PropertyInterface
-from ..common_interface import LowerLimitSaturationIronInterface
-from ..common_interface import LowerLimitSaturationChromiumInterface
-from ..common_interface import LowerLimitSaturationNickelInterface
-from ..common_interface import LowerLimitSaturationSiliconInterface
-from ..common_interface import LowerLimitSaturationAluminiumInterface
-from ..common_interface import LowerLimitChromiumInterface
-from ..common_interface import LowerLimitNickelInterface
-from ..common_interface import LowerLimitIronInterface
+from ..tch_common_interface import LowerLimitSaturationIronInterface
+from ..tch_common_interface import LowerLimitSaturationChromiumInterface
+from ..tch_common_interface import LowerLimitSaturationNickelInterface
+from ..tch_common_interface import LowerLimitSaturationSiliconInterface
+from ..tch_common_interface import LowerLimitSaturationAluminiumInterface
+from ..tch_common_interface import LowerLimitChromiumInterface
+from ..tch_common_interface import LowerLimitNickelInterface
+from ..tch_common_interface import LowerLimitIronInterface
 from .solubility_in_lead import OxygenSolubility
 from .solubility_in_lead import ChromiumSolubilityGosse2014
 from .solubility_in_lead import ChromiumSolubilityAlden1958
