+++ lbh15/properties/lbe_properties.py
@@ -4,13 +4,23 @@ from typing import List
 from typing import Union
 import numpy as np
 from scipy.constants import atm
-from .interface import PropertyInterface
+from .tph_common_interface import SaturationVapourPressureInterface
+from .tph_common_interface import SurfaceTensionInterface
+from .tph_common_interface import DensityInterface
+from .tph_common_interface import ThermalExpansionInterface
+from .tph_common_interface import SpeedOfSoundInterface
+from .tph_common_interface import IsentropicCompressibilityInterface
+from .tph_common_interface import SpecificHeatInterface
+from .tph_common_interface import SpecificEnthalpyInterface
+from .tph_common_interface import DynamicViscosityInterface
+from .tph_common_interface import ElectricalResistivityInterface
+from .tph_common_interface import ThermalConductivityInterface
 from .._decorators import range_warning
 from .._commons import LBE_MELTING_TEMPERATURE as T_m0
 from .._commons import LBE_BOILING_TEMPERATURE as T_b0
 
 
-class p_s(PropertyInterface):
+class p_s(SaturationVapourPressureInterface):
     """
     Liquid lead-bismuth eutectic *saturation vapour pressure* property class.
     """
@@ -78,20 +88,6 @@ class p_s(PropertyInterface):
         return [T_m0, T_b0]
 
     @property
-    def units(self) -> str:
-        """
-        str : Saturation vapour pressure unit
-        """
-        return "[Pa]"
-
-    @property
-    def long_name(self) -> str:
-        """
-        str : Saturation vapour pressure long name
-        """
-        return "saturation vapour pressure"
-
-    @property
     def description(self) -> str:
         """
         str : Saturation vapour pressure description
@@ -99,7 +95,7 @@ class p_s(PropertyInterface):
         return f"Liquid lbe {self.long_name}"
 
 
-class sigma(PropertyInterface):
+class sigma(SurfaceTensionInterface):
     """
     Liquid lead-bismuth eutectic *surface tension* property class.
     """
@@ -144,20 +140,6 @@ class sigma(PropertyInterface):
         return [T_m0, 1400.0]
 
     @property
-    def units(self) -> str:
-        """
-        str : Surface tension unit
-        """
-        return "[N/m]"
-
-    @property
-    def long_name(self) -> str:
-        """
-        str : Surface tension long name
-        """
-        return "surface tension"
-
-    @property
     def description(self) -> str:
         """
         str : Surface tension description
@@ -165,7 +147,7 @@ class sigma(PropertyInterface):
         return f"Liquid lbe {self.long_name}"
 
 
-class rho(PropertyInterface):
+class rho(DensityInterface):
     """
     Liquid lead-bismuth eutectic *density* property class.
     """
@@ -209,20 +191,6 @@ class rho(PropertyInterface):
         return [T_m0, T_b0]
 
     @property
-    def units(self) -> str:
-        """
-        str : Density unit
-        """
-        return "[kg/m^3]"
-
-    @property
-    def long_name(self) -> str:
-        """
-        str : Density long name
-        """
-        return "density"
-
-    @property
     def description(self) -> str:
         """
         str : Density description
@@ -230,7 +198,7 @@ class rho(PropertyInterface):
         return f"Liquid lbe {self.long_name}"
 
 
-class alpha(PropertyInterface):
+class alpha(ThermalExpansionInterface):
     """
     Liquid lead-bismuth eutectic *thermal expansion coefficient*
     property class.
@@ -269,20 +237,6 @@ class alpha(PropertyInterface):
         return [T_m0, T_b0]
 
     @property
-    def units(self) -> str:
-        """
-        str : Thermal expansion coefficient unit
-        """
-        return "[1/K]"
-
-    @property
[…]
