### 2024-03-11 b87c65e
[160_common_interface], issue #160, updated sources for documentation (defined tph*.rst and tch*.rst) +  minor fixes in tch_common_interface docstrings
