+++ docs/source/tch_common_interface.rst
@@ -1,0 +1,7 @@
+*tch_common_interface* Module
+=============================
+
+.. automodule:: lbh15.properties.tch_common_interface
+    :members:
+    :member-order: bysource
+    :inherited-members:
