+++ docs/source/properties.rst
@@ -13,13 +13,15 @@ These classes represent the properties dynamically loaded by the :class:`.Lead`,
 
    interface.rst
 
-   common_interface.rst
+   tph_common_interface.rst
 
    lead_properties.rst
 
    bismuth_properties.rst
 
    lbe_properties.rst
+
+   tch_common_interface.rst
 
    lead_thermochemical_properties.rst
 
