+++ docs/source/tph_common_interface.rst
@@ -1,0 +1,7 @@
+*tph_common_interface* Module
+=============================
+
+.. automodule:: lbh15.properties.tph_common_interface
+    :members:
+    :member-order: bysource
+    :inherited-members:
