### 2024-03-17 56a75dc
[148_statement_of_need], issue - #148, Refactored the layout of the first two paragraphs of the introduction of README and doc
