+++ docs/source/readme_full.rst
@@ -10,9 +10,9 @@ Introduction
 thermo-physical and the thermo-chemical properties of lead, bismuth and lead-bismuth eutectic (lbe) metal alloy available from
 the handbook edited by OECD/NEA :cite:`Agency2015`:
 https://www.oecd-nea.org/jcms/pl_14972/handbook-on-lead-bismuth-eutectic-alloy-and-lead-properties-materials-compatibility-thermal-hydraulics-and-technologies-2015-edition.
-This package offers a standard implementation for use of these properties in scientific applications.
 
-*lbh15* is used for numerical simulation and experimental activities supporting the
+*lbh15* offers a standard implementation for use of these properties in scientific applications.
+In particular, this package is used for numerical simulation and experimental activities supporting the
 development of nuclear reactors that employ heavy liquid metals as primary coolant.
 
 The thermo-physical properties implemented in the package are listed in :numref:`tableprop`.
