### 2024-03-17 ff4deac
[150_contribution], issue - #150, Added location and framework of tests in the contributing doc
