### 2024-03-17 113ed99
[141_joss_paper_reviewer_2], issue - #141, Corrected some typos in the modifications
