### 2024-03-17 bdc1abc
[145_joss_paper_reviewer_1], issue - #145, Implemented modifications related to Statement of Need section
