### 2024-03-17 874d9b6
[145_joss_paper_reviewer_1], issue - #145, Implemented modifications related to Implementation section
