+++ joss/paper.bib
@@ -24,6 +24,17 @@
   publisher = {ANS}
 }
 
+@software{iapwscode,
+  author       = {Romera, J. J. G.},
+  title        = {jjgomera/iapws:},
+  month        = may,
+  year         = 2021,
+  publisher    = {Zenodo},
+  version      = {v1.5.2},
+  doi          = {10.5281/zenodo.4744318},
+  url          = {https://doi.org/10.5281/zenodo.4744318}
+}
+
 @techreport{iapws,
   author = {{International Association for the Properties of Water and Steam}},
   title = {Revised Release on the IAPWS Formulation 1995 for the Thermodynamic Properties of Ordinary Water Substance for General and Scientific Use},
