### 2024-03-17 c008bc4
[145_joss_paper_reviewer_1], issue - #145, Implemented modifications related to Use section
