### 2024-03-17 23408ca
[145_joss_paper_reviewer_1], issue - #145, Implemented modifications related to Implemented Properties section
