### 2024-03-17 efe8232
[145_joss_paper_reviewer_1], issue - #145, Implemented modifications related to Implementation History section
