### 2024-03-17 120f7b5
[145_joss_paper_reviewer_1], issue - #145, Implemented modifications related to Documentation section
