+++ joss/paper.md
@@ -140,8 +140,7 @@ The documentation of `lbh15` is generated by `Sphinx` and published on
 `https://newcleo-dev-team.github.io/lbh15/index.html`.
 
 It is composed of parts addressed separately to the developers and to the 
-users. An advanced use of the package needs skills in Python software 
-development. The documentation contains examples for users, 
+users. The documentation contains examples for users, 
 from basic use to short tutorials for more advanced 
 applications.
 
