### 2024-03-19 4a7fd55
[145_joss_paper_reviewer_1], issue - #145, Applied a few style modifications
