### 2024-03-19 newcleo-dev-team/lbh15#141 @LukeSeifert
### Quality of writing
@lelaus The paper is looking good! I just have a few more comments on minor things. Feel free to @ me once you have addressed these, and I think I can go ahead and close this issue.

#### Summary
- Line 15: "As alternative" -> "Alternatively"
- Line 16: "At liquid metal object's" -> "At the liquid metal object's"

#### Statement of Need
- Line 26: "As well" -> "Additionally"
- Line 30: Comma after "system"?

#### Implementation
- Line 38: I think a semicolon would be better than a colon here (suggestion).

#### Use
- Line 58: Add "and" before "Prandtl number" to match with the "thermo-chemical" bullet point
