### 2024-03-19 newcleo-dev-team/lbh15#140 @LukeSeifert
Great changes @lelaus! My only remaining issue here is that the readme "How to Contribute" link doesn't currently work for me (links to https://github.com/newcleo-dev-team/lbh15/blob/master/CONTRIBUTIONS instead of https://github.com/newcleo-dev-team/lbh15/blob/master/CONTRIBUTIONS.rst).

Once you have fixed this, @ me and I will close this issue.
