### 2024-03-19 790ed7f
[141_joss_paper_reviewer_2], issue - #141, Applied some minor modifications as suggested by the reviewer
