### 2024-03-20 newcleo-dev-team/lbh15#140 @lelaus
The error you noticed have been fixed.
Thanks, @LukeSeifert for your valuable analysis and comments!
