### 2024-03-20 6599f76
[158_remove_page_repetition], issue - #158, Removed from doc documentation.rst and modified lbh15.rst accordingly
