### 2024-03-21 40aa211
[159_properties_monospace], issue - #159, Converted property names into monospace
