+++ docs/source/lead.rst
@@ -25,138 +25,138 @@ In detail, a :class:`.Lead` object comes with the following default properties:
 
   a. **Thermo-physical**:
 
-    - :math:`T_{m0}` melting temperature :math:`\left[K\right]`:
+    - ``T_m0`` melting temperature :math:`\left[K\right]`:
 
         :math:`600.6`
-    - :math:`Q_{m0}` melting latent heat :math:`\left[\frac{J}{kg}\right]`:
+    - ``Q_m0`` melting latent heat :math:`\left[\frac{J}{kg}\right]`:
 
         :math:`23.07 \cdot 10^3`
-    - :math:`T_{b0}` boiling temperature :math:`\left[K\right]`:
+    - ``T_b0`` boiling temperature :math:`\left[K\right]`:
 
         :math:`2021`
-    - :math:`Q_{b0}` vaporisation heat :math:`\left[\frac{J}{kg}\right]`:
+    - ``Q_b0`` vaporisation heat :math:`\left[\frac{J}{kg}\right]`:
 
         :math:`858.6 \cdot 10^3`
-    - :math:`p_s` saturation vapour pressure :math:`\left[Pa\right]`:
+    - ``p_s`` saturation vapour pressure :math:`\left[Pa\right]`:
 
         :math:`p_s\left(T\right) = \displaystyle 5.76\cdot10^9\cdot\exp{\left(-22131/T\right)}`
-    - :math:`\sigma` surface tension :math:`\left[\frac{N}{m}\right]`:
+    - ``sigma`` surface tension :math:`\left[\frac{N}{m}\right]`:
 
         :math:`\sigma\left(T\right) = \displaystyle\left(525.9 - 0.113{\cdot}T\right)\cdot10^{-3}`
-    - :math:`u_s` speed of sound :math:`\left[\frac{m}{s}\right]`:
+    - ``u_s`` speed of sound :math:`\left[\frac{m}{s}\right]`:
 
         :math:`u_s\left(T\right) = \displaystyle 1953 - 0.246{\cdot}T`
-    - :math:`\alpha` thermal expansion coefficient :math:`\left[\frac{1}{K}\right]`:
+    - ``alpha`` thermal expansion coefficient :math:`\left[\frac{1}{K}\right]`:
 
         :math:`\alpha\left(T\right) = \displaystyle \left(8942 - T\right)^{-1}`
-    - :math:`c_p` specific heat capacity :math:`\left[\frac{J}{kg{\cdot}K}\right]`:
+    - ``cp`` specific heat capacity :math:`\left[\frac{J}{kg{\cdot}K}\right]`:
 
         :math:`c_p\left(T\right) = \displaystyle176.2 - 4.923\cdot10^{-2}{\cdot}T + 1.544\cdot10^{-5}{\cdot}T^2 - 1.524\cdot10^{6}{\cdot}T^{-2}`
-    - :math:`\rho` density :math:`\left[\frac{kg}{m^3}\right]`:
+    - ``rho`` density :math:`\left[\frac{kg}{m^3}\right]`:
 
         :math:`\rho_0\left(T\right) = \displaystyle11441 - 1.2795{\cdot}T`
 
         :math:`\rho\left(T,p\right) = \displaystyle\rho_0\left(T\right) + \left( \frac{1}{u_s\left(T\right)^2} + T \cdot \frac{\alpha\left(T\right)^2}{c_p\left(T\right)} \right) \cdot \left( p - p_{atm} \right)`, :math:`\quad` where:
         
         :math:`\quad p_{atm}=101325.00 \left[Pa\right]`
-    - :math:`\beta_s` isentropic compressibility :math:`\left[\frac{1}{Pa}\right]`:
[…]
