+++ docs/source/lbe.rst
@@ -25,132 +25,132 @@ In detail, a :class:`.LBE` object comes with the following default properties:
 
   a. **Thermo-physical**:
 
-    - :math:`T_{m0}` melting temperature :math:`\left[K\right]`:
+    - ``T_m0`` melting temperature :math:`\left[K\right]`:
 
         :math:`398.0`
-    - :math:`Q_{m0}` melting latent heat :math:`\left[\frac{J}{kg}\right]`:
+    - ``Q_m0`` melting latent heat :math:`\left[\frac{J}{kg}\right]`:
 
         :math:`38.6 \cdot 10^3`
-    - :math:`T_{b0}` boiling temperature :math:`\left[K\right]`:
+    - ``T_b0`` boiling temperature :math:`\left[K\right]`:
 
         :math:`1927`
-    - :math:`Q_{b0}` vaporisation heat :math:`\left[\frac{J}{kg}\right]`:
+    - ``Q_b0`` vaporisation heat :math:`\left[\frac{J}{kg}\right]`:
 
         :math:`856.6 \cdot 10^3`
-    - :math:`p_s` saturation vapour pressure :math:`\left[Pa\right]`:
+    - ``p_s`` saturation vapour pressure :math:`\left[Pa\right]`:
 
         :math:`p_s\left(T\right) = \displaystyle1.22 \cdot 10^{10} \cdot \exp{\left(-22552 / T\right)}`
-    - :math:`\sigma` surface tension :math:`\left[\frac{N}{m}\right]`:
+    - ``sigma`` surface tension :math:`\left[\frac{N}{m}\right]`:
 
         :math:`\sigma\left(T\right) = \displaystyle\left(448.5 - 0.0799 \cdot T\right) \cdot 10^{-3}`
-    - :math:`u_s` speed of sound :math:`\left[\frac{m}{s}\right]`:
+    - ``u_s`` speed of sound :math:`\left[\frac{m}{s}\right]`:
 
         :math:`u_s\left(T\right) = \displaystyle1855 - 0.212 \cdot T`
-    - :math:`\alpha` thermal expansion coefficient :math:`\left[\frac{1}{K}\right]`:
+    - ``alpha`` thermal expansion coefficient :math:`\left[\frac{1}{K}\right]`:
 
         :math:`\alpha\left(T\right) = \displaystyle\left(8558 - T\right)^{-1}`
-    - :math:`c_p` specific heat capacity :math:`\left[\frac{J}{kg{\cdot}K}\right]`:
+    - ``cp`` specific heat capacity :math:`\left[\frac{J}{kg{\cdot}K}\right]`:
 
         :math:`c_p\left(T\right) = \displaystyle164.8 - 3.94 \cdot 10^{-2} \cdot T + 1.25 \cdot 10^{-5} \cdot T^2 - 4.56 \cdot 10^5 \cdot T^{-2}`
-    - :math:`\rho` density :math:`\left[\frac{kg}{m^3}\right]`:
+    - ``rho`` density :math:`\left[\frac{kg}{m^3}\right]`:
 
         :math:`\rho_0\left(T\right) = \displaystyle11065 - 1.293 \cdot T`
 
         :math:`\rho\left(T,p\right) = \displaystyle\rho_0\left(T\right) + \left( \frac{1}{u_s\left(T\right)^2} + T \cdot \frac{\alpha\left(T\right)^2}{c_p\left(T\right)} \right) \cdot \left( p - p_{atm} \right)`, :math:`\quad` where:
         
         :math:`\quad p_{atm}=101325.00 \left[Pa\right]`
-    - :math:`\beta_s` isentropic compressibility :math:`\left[\frac{1}{Pa}\right]`:
[…]
