### 2024-03-22 newcleo-dev-team/lbh15#145 @lelaus
Dear @damar-wicaksono ,

we have tried to apply all your comments, remarks and suggestions. Please, find below our specific actions for each of your points.

In case of any misunderstanding and/or further comments and/or doubt, please do not hesitate to write us.

**Summary**
. Line 14: OK

. Lines 15-16:
	The authors agree with the reviewer's comment. Sentence changed into:
		_As alternative, the physical properties can be used at liquid metal object's instantiation, ..._
	See issue #141

**Statement of Need**
. Lines 28-29:
	changed _physical correlations_ into _correlations of physical properties_
	the "computational tools" have been specified by adding _, such as CFD, system and sub-channel codes concerning heavy liquid metals._

. Lines 25-26: OK

**Implementation**
. Line 31:
	Good point! Thanks!

. Line 34: OK

**Use**
. Line 46: OK

. Line 48: OK

. Line 48:
	The authors agree with the reviewer that the sentence is not clear.
	The sentence has been rewritten when managing issue #141 into:
		_In addition, this approach allows to select the default correlations of the properties by means of the available class methods._

. Line 51:
	The authors agree with the reviewer that the sentence is not clear.
	The second part of the sentence has been rewritten when managing issue #141 into:
		_... since it offers faster instantiation and evaluation of the correlation functions._

**Implemented Properties**
. These two subdivisions are not structured in parallel:
	The authors agree with the reviewer.
	The ":" has been inserted into both the subdivisions when managing issue #141, in order to make them parallel;

. I see no reason to capitalize the term "Oxygen".
	OK

**Implementation History**
. Line 66:
	_It improves the performance and readability._
	has been substituted with _It improves the overall understanding._

. Line 67:
	The point of view of the authors is:
	. The **vectorization** should improve the **usability** since the correlation functions may be called either by passing a single input (thus retrieving a single result) or by passing an array of inputs (thus retrieving an array of results at the same time). Without vectorization, the user should call the correlation functions multiple times, once for each element of the input array.
	. The **Horner scheme** is an efficient way of writing polynomials according to the provided reference: in this sense, coding polynomials according to this scheme improves the **performance** of the code.

. Line 68:
	"It" has been substituted with _The documentation of the current version_.

. Lines 71-72:
	The authors agree with the reviewer: "readability" has been removed since it would be a repetition, and the code quality sentences are made more concise and are moved upperwards.

. pycodestyle and pylint: OK

. Lines 72-74: OK

**Documentation**
. Line 82:
	The authors agree with the reviewer, thus the sentence has been removed.

. Line 83:
	"is completed with" has been substituted with "contains" when handling issue #141.
	The authors think that the modification complies also with what stated by the reviewer.

. Line 83:
	. "users" instead of "the users":
		already applied when handling issue #141
	. "starting from...":
		already applied when handling issue #141
	. here "_basic use_" and "_advanced applications_" are considered as general expressions, thus not needing a strict reference to the chapter naming adopted in the doc.
