### 2024-03-22 newcleo-dev-team/lbh15#145 @damar-wicaksono
Hello @lelaus,

I have read the updated draft manuscript and can confirmed the changes you made; along with the suggestions from the other reviewer, I think it is now a better read.

Thank you very much for taking the time going through my comments, addressing them, and providing explanations to some of my questions!

A couple other very minor suggestions:

- Line 32: Maybe add a space between `iapws` and "(Romera, 2021)"?
- As suggested by the editorialbot, it appears your own publication at NURETH has its [DOI:10.13182/nureth20-40554](https://doi.org/10.13182/nureth20-40554). Maybe consider adding the DOI to the manuscript?

Beyond these additional minor suggestions, I have no further comments regarding the manuscript. Feel free to close this issue afterwards (if changes are made). Thank you!
