### 2024-03-22 58c1a33
[145_joss_paper_reviewer_1], issue - #145, Applied new modifications as per reviewer indications
