+++ joss/paper.bib
@@ -21,7 +21,8 @@
   year      =  {2023},
   pages     = {1--12},
   month = {Aug 20--25},
-  publisher = {ANS}
+  publisher = {ANS},
+  doi       = {10.13182/nureth20-40554}
 }
 
 @software{iapwscode,
