### 2024-03-22 61cae80
[144_joss_functionality_reviewer_1], issue - #144, Corrected typo about the name of method available_correlations()
