### 2024-03-22 1e77498
[156_refactor_advanced_usage], issue - #156, Corrected typo referred to task n.3
