### 2024-03-24 350ef67
[145_joss_paper_reviewer_1], issue - #145, Added missing DOI
