+++ joss/paper.bib
@@ -1,9 +1,11 @@
 @book{nea,
-  title = {Handbook on Lead-Bismuth Eutectic Alloy and Lead Properties, Materials Compatibility, Thermal-Hydraulics and Technologies},
-  author = {Fazio, C. et al.},
-  year = {2015},
+  title       = {Handbook on Lead-Bismuth Eutectic Alloy and Lead Properties, Materials Compatibility, Thermal-Hydraulics and Technologies},
+  author      = {Fazio, C. et al.},
+  year        = {2015},
   institution = {OECD/Nuclear Energy Agency (NEA), Paris, France},
-  number = {No. 7268}
+  number      = {No. 7268},
+  url         = {https://www.oecd-ilibrary.org/content/publication/42dcd531-en},
+  doi         = {10.1787/42dcd531-en}
 }
 
 @book{pybook,
@@ -20,7 +22,7 @@
   booktitle = {Proceedings of 20th International Topical Meeting on Nuclear Reactor Thermal Hydraulics (NURETH-20), Washington DC, USA},
   year      =  {2023},
   pages     = {1--12},
-  month = {Aug 20--25},
+  month     = {Aug 20--25},
   publisher = {ANS},
   doi       = {10.13182/nureth20-40554}
 }
@@ -32,20 +34,20 @@
   year         = 2021,
   publisher    = {Zenodo},
   version      = {v1.5.2},
-  doi          = {10.5281/zenodo.4744318},
-  url          = {https://doi.org/10.5281/zenodo.4744318}
+  url          = {https://doi.org/10.5281/zenodo.4744318},
+  doi          = {10.5281/zenodo.4744318}
 }
 
 @techreport{iapws,
   author = {{International Association for the Properties of Water and Steam}},
-  title = {Revised Release on the IAPWS Formulation 1995 for the Thermodynamic Properties of Ordinary Water Substance for General and Scientific Use},
-  year = {2018},
+  title  = {Revised Release on the IAPWS Formulation 1995 for the Thermodynamic Properties of Ordinary Water Substance for General and Scientific Use},
+  year   = {2018},
   number = {IAPWS R6-95(2018)}
 }
 
 @book{hornerbook,
-  author = {Hildebrand, F. B.},
-  title = {Introduction to Numerical Analysis},
+  author    = {Hildebrand, F. B.},
+  title     = {Introduction to Numerical Analysis},
   publisher = {McGraw-Hill},
-  year = {1974}
+  year      = {1974}
 }
