### 2024-03-24 newcleo-dev-team/lbh15#145 @lelaus
Yes, thanks @damar-wicaksono for noting this. I verified it is ok. About the other references I did not find any DOI.
