### 2024-03-24 ff0bbc0
[156_refactor_advanced_usage], issue - #156, Extracted available_correlations explanation outside from bulleted list in advanced usage chapter
