+++ docs/source/readme_full.rst
@@ -588,7 +588,32 @@ In the following, some examples are provided:
 Advanced Usage
 ++++++++++++++
 
-Advanced usage includes the possibility of adding new properties and new physical correlations according to the following examples.
+Advanced usage includes the possibility of adding new properties and new physical
+correlations. First of all, the user should be aware of the properties that are
+currently available along with the correlations. The method
+:meth:`.Lead.available_correlations` provides this information for a
+:class:`.Lead` instance (the same goes for :class:`.Bismuth` and :class:`.LBE`
+instances). Two examples of use are provided, the first one asking for the
+available correlations for all the implemented properties, the second one
+asking for the available correlations for a few specific properties, that is,
+the heat capacity and the speed of sound:
+
+>>> from lbh15 import Lead
+>>> Lead.available_correlations()
+defaultdict(<class 'list'>, {'k': ['lbh15'], 'lim_ni': ['lbh15'], 'u_s': ['sobolev2011'], 'mu': ['lbh15'], 'H': ['lbh15'], 'lim_cr': ['venkatraman1988', 'alden1958', 'gosse2014'], 'sigma': ['jauch1986'], 'cp': ['sobolev2011', 'gurvich1991'], 'cr_sol': ['venkatraman1988', 'gosse2014', 'alden1958'], 'in_dif': ['lbh15'], 'o_pp': ['taskinen1979', 'charle1976', 'alcock1964', 'otsuka1979', 'otsuka1981', 'fisher1966', 'isecke1977', 'szwarc1972', 'ganesan2006'], 'lim_cr_sat': ['lbh15'], 'si_sol': ['lbh15'], 'fe_dif': ['lbh15'], 'lim_si': ['lbh15'], 'o_dif': ['charle1976', 'swzarc1972', 'arcella1968', 'gromov1996', 'ganesan2006b', 'otsuka1975', 'homna1971'], 'G': ['lbh15'], 'p_s': ['sobolev2011'], 'beta_s': ['lbh15'], 'se_dif': ['lbh15'], 'lim_al_sat': ['lbh15'], 'ni_sol': ['gosse2014'], 'r': ['lbh15'], 'S': ['lbh15'], 'h': ['sobolev2011'], 'alpha': ['lbh15'], 'co_dif': ['lbh15'], 'lim_si_sat': ['lbh15'], 'lim_fe_sat': ['lbh15'], 'fe_sol': ['gosse2014'], 'lim_fe': ['lbh15'], 'rho': ['sobolev2008a'], 'te_dif': ['lbh15'], 'lim_ni_sat': ['lbh15'], 'o_sol': ['lbh15']})
+>>> Lead.available_correlations(["cp", "u_s"])
+{'cp': ['gurvich1991', 'sobolev2011'], 'u_s': ['sobolev2011']}
+
+In the first example, the keys of the returned dictionary represents all the
+properties currently available, and, for each one, a list is provided containing
+the names of all the related correlations. The name :code:`lbh15` is used when
+only the default correlation is available. In the second example, the structure
+of the returned dictionary is the same, but containing only data for the
+required properties.
+
+The rest of this chapter is divided into two sections, the first describing
+how to add a new correlation to an existing property, the second describing
+how to add a new property with its correlation.
 
 - This first example shows how to define a custom correlation for the liquid lead density that 
   does not depend on pressure, and how to use it instead of the default one (the same holds for :class:`.Bismuth` and :class:`.LBE` classes).
