### 2024-03-24 31c3cca
[156_refactor_advanced_usage], issue - #156, Created two subsections into Advanced Usage section
