### 2024-03-24 341a185
[156_refactor_advanced_usage], issue - #156, Inserted example for default correlation choice when adding new correlation
