### 2024-03-25 9919a28
[156_refactor_advanced_usage], issue - #156, Corrected a few typos into advanced usage section
