### 2024-03-25 2265408
[153_check_separator], added replace of \ with / in POSIX to replicate the behaviour of os.path.normpath in Windows
