### 2024-03-25 989e429
[157_doc_validity_range], issue #157, rephrased the last sentence in ranges of validity
