### 2024-03-26 d8a8b1a
[133_paper_joss], issue - #133, Converted English words into the corresponding American ones
