### 2024-03-26 efe6173
[147_joss_documentation_reviewer_1], issue - #147, Homogenized all the words to US spelling
