### 2024-03-27 428c274
[147_joss_documentation_reviewer_1], issue - #147, Managed task related to Initialization from Properties section
