### 2024-03-27 ff1fbeb
[147_joss_documentation_reviewer_1], issue - #147, Changed elements' name from uppercase to lowercase
