+++ docs/source/learn_more.rst
@@ -4,17 +4,17 @@
 Oxygen Control
 ++++++++++++++
 
-In lead and LBE systems, Oxygen is the most important chemical element, which results
+In lead and LBE systems, oxygen is the most important chemical element, which results
 from start-up operations, maintenance services and, possibly, incidental contaminations (:cite:`Agency2015`).
 To operate a nuclear reactor cooled by a lead alloy, it is thus important to
-determine the upper and the lower Oxygen concentration limits.
+determine the upper and the lower oxygen concentration limits.
 
 .. _ Oxygen concentration upper limit:
 
 Oxygen Concentration Upper Limit
 ================================
 
-The upper limit corresponds to the Oxygen concentration value above which contamination by coolant oxides occurs.
+The upper limit corresponds to the oxygen concentration value above which contamination by coolant oxides occurs.
 It is represented by the *Oxygen solubility* in lead and LBE alloys. *lbh15* provides
 these properties in the ``lead_thermochemical_properties.solubility_in_lead``
 and ``lbe_thermochemical_properties.solubility_in_lbe`` modules.
@@ -31,17 +31,17 @@ The lower limit corresponds to the minimum value of the oxygen concentration ena
 The oxide layer formation is possible only when the oxygen potential in the liquid metal is above the
 potential leading to the protective film formation. The correlations implemented in the
 ``lead_thermochemical_properties.lead_oxygen_limits`` and ``lbe_thermochemical_properties.lbe_oxygen_limits``
-modules for computing the lower limits of Oxygen concentration are obtained by applying the methodology
+modules for computing the lower limits of oxygen concentration are obtained by applying the methodology
 described in :cite:`Agency2015`, chapter 4, part 4.2.2, pages 187-192. A brief summary is provided in the following.
 
-First of all, the reference reaction equation and the associated Gibbs free energy are determined. Then, the Oxygen
+First of all, the reference reaction equation and the associated Gibbs free energy are determined. Then, the oxygen
 concentration is expressed as a function of temperature. Eventually, two kinds of correlations, based on two different
 assumptions, are derived.
 
 ..
 
   1. The equation :eq:`oxide_react_eq` of the oxidation reaction is set by considering that it occurs
-     between the metal and the Oxygen, with the Oxygen supposed in solution as dissolved :math:`\ce{PbO}` below its saturation limit.
+     between the metal and the oxygen, with the oxygen supposed in solution as dissolved :math:`\ce{PbO}` below its saturation limit.
      The formation equation :eq:`metal_oxide_eq` of the metal oxide (equation 4.5, page 188 of :cite:`Agency2015`) is combined with the formation
      equation :eq:`pbo_eq` of :math:`\ce{PbO}`, (table 4.2.2, page 189 of :cite:`Agency2015`):
 
@@ -121,22 +121,22 @@ assumptions, are derived.
 ..
 
   4. The aim is now to develop, for each possible dissolved metal, a correlation for the lower limit of the
-     Oxygen concentration that has the same structure as the equation 4.12, part 4.2.2 of :cite:`Agency2015`. Starting from the
+     oxygen concentration that has the same structure as the equation 4.12, part 4.2.2 of :cite:`Agency2015`. Starting from the
      oxidation reaction equation :eq:`oxide_react_eq`, the following substitution is applied:
 
[…]

[…]
