### 2024-03-27 1a1c309
[147_joss_documentation_reviewer_1], issue - #147, Applied modifications suggested for the Ranges of Validity section
