+++ docs/source/learn_more.rst
@@ -4,7 +4,7 @@
 Oxygen Control
 ++++++++++++++
 
-In lead and LBE systems, oxygen is the most important chemical element, which results
+In lead and lbe systems, oxygen is the most important chemical element, which results
 from start-up operations, maintenance services and, possibly, incidental contaminations (:cite:`Agency2015`).
 To operate a nuclear reactor cooled by a lead alloy, it is thus important to
 determine the upper and the lower oxygen concentration limits.
@@ -15,7 +15,7 @@ Oxygen Concentration Upper Limit
 ================================
 
 The upper limit corresponds to the oxygen concentration value above which contamination by coolant oxides occurs.
-It is represented by the *Oxygen solubility* in lead and LBE alloys. *lbh15* provides
+It is represented by the *oxygen solubility* in lead and lbe alloys. *lbh15* provides
 these properties in the ``lead_thermochemical_properties.solubility_in_lead``
 and ``lbe_thermochemical_properties.solubility_in_lbe`` modules.
 The implemented data are extracted from :cite:`Agency2015`, table 3.5.2,
@@ -115,7 +115,7 @@ assumptions, are derived.
      
      .. math:: \alpha_i=C_i / C_i^{sat}.
 
-     About the chemical activity of lead in LBE, *lbh15* implements the correlation proposed by Gossé (2014)
+     About the chemical activity of lead in lbe, *lbh15* implements the correlation proposed by Gossé (2014)
      and written in chapter 3.3, part 3.3 of :cite:`Agency2015`.
 
 ..
@@ -141,8 +141,8 @@ assumptions, are derived.
      a useful correlation, the user can choose between two strategies that are proposed and adopted in *lbh15*. They differ
      on how they treat the chemical activity of the dissolved metal. The actual activities at the interface are
      influenced by how diffusion, convection and mass transfer phenomena interact in the liquid metal boundary layer.
-     Ongoing researches are in progress, but currently the exact values for the chemical activities of the dissolved
-     metal and of the Oxygen are not known.
+     In spite of the ongoing research, the exact values for the chemical activities of the dissolved
+     metal and of the oxygen are not known.
 
      a. The first approach is to consider the chemical activity of the dissolved metal equal to one.
         In this way, the first and the second terms of the right hand side of equation :eq:`ox_conc_eq` become zero, enabling to
@@ -156,7 +156,7 @@ assumptions, are derived.
 
         - :math:`C_{\ce{O}}^{sat}` is computed by adopting the recommended coefficients from table 3.5.2 of :cite:`Agency2015`;
 
-        - :math:`\alpha_{\ce{Pb}}` is taken equal to one in pure Lead, while in LBE it is computed by adopting the
+        - :math:`\alpha_{\ce{Pb}}` is taken equal to one in pure Lead, while in lbe it is computed by adopting the
           correlation proposed by Gossé as indicated at page 146 of :cite:`Agency2015`.
 
      b. The second approach does not exploit any assumption. In order to make equation :eq:`ox_conc_eq` solvable, the two unknowns
@@ -178,9 +178,9 @@ Ranges of Validity
 ==================
 
 As stated in the previous section, multiple correlations are involved in the computation of the lower limits of
-oxygen concentration, each being valid over a specific temperature range. The temperature range of
-validity specified in *lbh15* for each lower limit is given by the intersection of the
-ranges of the correlations which the lower limit depends on. More details as follows:
+oxygen concentration, each being valid over a specific temperature range. The validity range of a specific lower limit
+is then defined as the intersection of the validity ranges of the correlations on which the lower limit itself depends.
+More details as follows:
 
   - For the lower limit correlations based on the saturation assumption (approach *a*), the lower temperature
     value is taken equal to the lower limit of the validity range of the oxygen solubility correlation,
[…]
