### 2024-03-27 66cf7c4
[147_joss_documentation_reviewer_1], issue - #147, Applied modifications related to the Tutorials section
