### 2024-03-27 1d92968
[147_joss_documentation_reviewer_1], issue - #147, Applied modifications related to lead, bismuth and lbe modules
