### 2024-03-27 newcleo-dev-team/lbh15#147 @lelaus
Dear @damar-wicaksono , we are working for applying the modifications you suggested.

About your remarks in the last section "Regarding the tests", we are not able to reproduce the errors you got. Could you please describe us the procedure you follow before getting the reported errors?

Thanks!
