### 2024-03-28 ba5c297
[153_check_separator], issue - #153, Used only pathlib methods into set_custom_property_path method
