### 2024-03-29 89c04a9
[153_check_separator], issue - #153, Improved set_custom_properties_path method
