### 2024-03-29 newcleo-dev-team/lbh15#144 @lelaus
Dear @damar-wicaksono ,

the authors agree about all the points you noticed and tried to manage them. In the following, you could find the corresponding details (each single task has been identified by a number starting from top):

1. Handled in #152

2. Handled in #154

3. As the reviewer already noted, this feature has been added in #136, which has been managed after the release of 2.0.0 version, thus not being available in the PyPI.
   The idea is to release the 2.1.0 version of lbh15 once terminated the JOSS review: this feature will be present in the 2.1.0 version, and the documentation will be updated accordingly.

4. Great finding! It is a typo. Corrected!

5. Handled in #153.
