### 2024-03-29 newcleo-dev-team/lbh15#147 @lelaus
Dear @damar-wicaksono ,

the authors agree with your comments and suggestions. They tried to manage all of them. The details of all the actions performed correspondingly is in the following:

**General remark: British English vs American English**
- 	Adopted American English;
- 	Corrected JOSS paper in issue #133 by converting English words _vapourization_ and _vectorisation_ into the American ones _vaporization_ and _vectorization_;
- 	Corrected documentation by homogenizing all terms to the US version;

**Introduction**
- task managed in issue #148 by better describing the statement of need in the introduction section of both the README and the documentation

**Installation**
- task managed in issue #155

**Basic Usage**
- `pprint` removed from the doc; task managed in issue #154

**Initialization from Properties**
- done

**Advanced Usage**
- Two new subsections have been created, that is, **How to Add New Correlations** and **How to Add New Properties**
- The description of how to retrieve the available correlations has been put outside from the bullet points (that have become two subsections); two examples have been provided, one for retrieving all the correlations, the other for retrieving the correlations of a few specified properties;
- done
- task managed in issue #153
- done
- done: it was a typo
- an example is added to show which is the _default behavior_
- done

**Learn More**
- done
- done
- done

**Oxygen Concentration Lower Limit**
- sentence rephrased into _In spite of the ongoing research, the exact ..._

**Ranges of Validity**
- sentence rephrased into _Such a validity range is defined as the intersection of the validity ranges of the correlations on which the corresponding lower limit depends_
- done
- any additional source cited in the documentation is already cited in the Handbook, thus we removed this citation
- the authors agree with the reviewer. _extension_ term is used erroneously, leading to misunderstandings. The authors would like to talk about the _amplitude of the validity range_. Thus the sentence has been rephrased into: 	_preference is given to those correlations whose range of validity is based on the lowest available temperature value and is the widest available_

**Tutorials**
- done
- done
- done
- done

**API Guide**
- removed the second of the two identified pages, the one without the title

**lead Module, bismuth Module, and lbe Module**
- done
- task managed in issue #159
- good point to avoid repetitions in code; the code has been refactored accordingly by handling issue #160

**Community Guidelines**
- task managed in issue #150
