### 2024-04-02 newcleo-dev-team/lbh15#144 @damar-wicaksono
Hi @lelaus,

I went through the issues I raised above and find that everything goes fine now. Thank you very much for handling the issues!

I'm closing this issue now.
