### 2024-04-02 newcleo-dev-team/lbh15#147 @damar-wicaksono
Hi @lelaus,

Thank you very much for considering and handling the issues I raised.

I went through the documentation once more and confirm that most issues are fixed, including the one related to the community guideline document.

There are however, a couple of minor issues that you said they are revised (meaning, I assume, you have no objections with) that I cannot confirm both in the online documentation and the source (see unchecked boxes above). Perhaps I was looking at the wrong docs/sources?
