### 2024-04-02 newcleo-dev-team/lbh15#147 @lelaus
Hi @damar-wicaksono ,

first of all, we would like to thank you very much for the accurate and thorough work you have done in reviewing our project. We are very happy as your analysis has contributed a lot to improving the quality of the work.

As for the "missing" tasks, I think that you didn't find the corresponding implemented actions on the master branch since the actions implemented in the current issue have not yet been merged. Now we have done it!

Please, let me know if there are any more problems.

Thanks!
