### 2024-04-02 newcleo-dev-team/lbh15#177 @damar-wicaksono
JOSS Submission Review (Reviewer 1): Clarify Authors Contributions
This issue is related to the [JOSS submission review](https://github.com/openjournals/joss-reviews/issues/6383) and would be my final suggestion regarding the paper.

Currently, the paper lists four authors, namely, G. Ottino, D. Panico, D. Tomatis, and P.-A. Pantel.
Looking at the commit contributions [page](https://github.com/newcleo-dev-team/lbh15/graphs/contributors), I've seen four contributors have made commits; are these the same people as the authors?
If not, I think it would be nice to clarify who did what related to the package/paper.
(Also, The `README.rst` list two authors and and three contributors).

If I understand things correctly, not every authors need to make a code contribution because the scope of contributions can be large.
Therefore, it would be nice for transparency reason to list the contributions of each authors separately.
I've seen many JOSS papers (for example, [here](https://joss.theoj.org/papers/10.21105/joss.05873)) adopted the [CRediT](https://credit.niso.org/) taxonomy to clarify those roles.

Perhaps the authors would consider doing the same or something similar?
