### 2024-04-03 060e10e
[177_credit_authors], issue - #177, Added authors' contributions
