### 2024-04-03 6f1c48a
[177_credit_authors], issue - #177, Modified style of authors' contributions
