### 2024-04-03 newcleo-dev-team/lbh15#177 @lelaus
Hi @damar-wicaksono ,

the authors agree with your suggestion.

We have added to the paper the authors' contribution section according to your indications. This modification has been merged to #133, which is dedicated to the JOSS paper.

Please, let us know if you have any additional comments.

Thanks!
