### 2024-04-03 newcleo-dev-team/lbh15#177 @damar-wicaksono
Hi @lelaus, thanks a lot for the prompt response. I confirm that you added the authors contributions section to the paper.

I'm closing this issue and will proceed to conclude the review of your nice package. Thank you again for handling all my comments and suggestions!
