### 2024-04-03 #6383 @damar-wicaksono
Hi @kellyrowland,

I've concluded my part of the review. I had several comments and suggestions about the paper, software, and its documentation that the authors/developers have satisfactorily addressed.

The package is well-designed, -implemented, and -documented; the paper is concise and well-written. I could imagine that the package would be very useful for anyone (especially from within the nuclear engineering communities) who needs a convenient, single-entry point access in Python to relevant physical properties data and correlations of heavy liquid metals which previously mostly only available to a wider audience in a "printed" form.

I'm happy to support the publication of this package in JOSS.
