### 2024-04-04 490f3ff
[179_pre_release_v210], issue - #179, Applied modifications as per setup.py guidelines point 1
