+++ docs/source/release_notes.rst
@@ -11,163 +11,50 @@ They are tracked as issues that have been managed during the development process
 three lists according to the labels assigned to them. To access the entire issue history and comments and commits
 related to each one, please look at https://github.com/newcleo-dev-team/lbh15.
 
+
 1. **Enhancements**
 
-   - Minor revisions to doc and comments - **#131**
+   - Increase abstraction level of `OxygenDiffusivityInterface` and `IronSolubility` classes - **#160**
 
-   - Remove typecheck for performance reasons - **#129**
+   - Including all the dependencies for doc building and testing into toml and `setup.py` - **#155**
 
-   - Review for v2.0.0 release - **#120**
-
-   - Tutorial: code and doc - **#118**
-
-   - Allow to deeply copy *LiquidMetal* objects - **#115**
-
-   - Need for a test covering all temperatures - **#113**
-
-   - :code:`set_custom_properties_path` Improvement - **#110**
-
-   - Making element-wise the correlation functions of all properties - **#103**
-
-   - Horner's Scheme - **#100**
-
-   - Update :code:`test_custom_properties` to test correct assignment of custom properties to different liquid metal classes - **#86**
-
-   - Thermo-chemical Modules Corrections and Improvements - **#76**
-
-   - Constants Into Commons - **#71**
-
-   - :code:`change_correlation_to_use()` Method (:code:`_lbh15.py`) Improvement - **#69**
-
-   - :code:`__assign_p()` Method (:code:`_lbh15.py`) Improvement - **#68**
-
-   - :code:`__compute_T()` Method (:code:`_lbh15.py`) Improvement - **#67**
-
-   - :code:`__check_temperature()` Method (:code:`_lbh15.py`) Improvement - **#66**
-
-   - :code:`__fill_class_attributes` Method (:code:`_lbh15.py`) Improvement - **#65**
-
-   - :code:`__check_properties()` Method (:code:`_lbh15.py`) Improvement - **#64**
-
-   - :code:`__add_property()` Method (:code:`_lbh15.py`) Improvement - **#63**
-
-   - :code:`__fill_instance_properties()` Method (:code:`_lbh15.py`) Improvement - **#62**
-
-   - Attribute Storing the Available Correlations (:code:`_lbh15.py`) - **#61**
-
-   - Attribute Storing All Property Objects (:code:`_lbh15.py`) - **#60**
-
-   - :code:`__load_custom_properties()` Method (:code:`_lbh15.py`) Improvement - **#59**
-
-   - :code:`__custom_properties_path` Attribute (:code:`_lbh15.py`) Modification - **#58**
-
-   - :code:`__property_list()` Method (:code:`_lbh15.py`) Improvement - **#57**
-
-   - :code:`_properties` Dict (:code:`_lbh15.py`) Keys Modification - **#56**
-
-   - :code:`is_injective()` Method (:code:`interface.py`) Name - **#54**
-
-   - :code:`info()` Method (:code:`interface.py`) Improvement - **#53**
-
-   - :code:`compute_bounds` Method (:code:`interface.py`) Improvement - **#52**
-
-   - Type Hints and Type Checking - **#49**
- 
-   - Define two new sections of the documentation to give the user more insights about thermo-chemical properties and their usage - **#48**
-
-   - Update the documentation to show the new thermo-chemical properties - **#47**
-
-   - Set the default correlations to use for the thermo-chemical properties of all liquid metals - **#44**
-
-   - Define module containing oxygen limits correlations related to lead - **#42**
-
-   - Define module containing thermo-chemical correlations related to lead - **#41**
-
-   - Define module containing diffusivity of elements in lead - **#40**
-
-   - Define module containing solubility of elements in lead - **#39**
-
-   - Define module containing oxygen limits correlations related to LBE - **#38**
-
[…]
