### 2024-04-04 #6383 @lelaus
Hi @kellyrowland ,
a new version of the software has been released including all the modifications made according to the reviewers' comments.

The latest release is **v2.1.0.**

The DOI of the corresponding archive on ZENODO is: **10.5281/zenodo.10927101**.

Please, let me know if you have any additional requests.

Thanks!
