### 2024-04-04 #6383 @kellyrowland
@lelaus could you please remove the formatting in the affiliation entries?
