### 2024-04-04 65e827b
[133_paper_joss], issue - #133, Removed formatting from affiliation name
