### 2024-04-04 #6383 @lelaus
Dear @kellyrowland ,

I have just applied this modification and pushed to the branch dedicated to JOSS paper.

Thanks!
