### 2024-04-04 #6383 @kyleniemeyer
Hello @lelaus, just a small edit before publishing: in the first paragraph of the Implementation History, can you change `described in [@nureth20]` to `described by @nureth20`? This will make it say "The release of version 1.1.0 of the package lbh15 was described by Panico and Tomatis (2023)."
