### 2024-04-04 5914c29
[133_paper_joss], issue - #133, Modified citation of nureth
